temp_scraper/requirements.txt
#requirements
requests
kafka-python
